% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.fusionTree}
\alias{as.fusionTree}
\title{Turn an hclust object into a fusion path.}
\usage{
as.fusionTree(hc_obj)
}
\arguments{
\item{hc_obj}{and hclust object}
}
\value{
a list with three element: path of fusion, vector of height, ordering.
}
\description{
Turn an hclust object into a fusion path.
}
