% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_method.R
\name{consensusTree}
\alias{consensusTree}
\title{Create a consensus tree.}
\usage{
consensusTree(lis, method = NULL, data_st = FALSE, dist_st = FALSE,
  tree_st = FALSE, data_center = TRUE, distance_method = "euclidean",
  linkage_method = "ward.D2", verbose = FALSE)
}
\arguments{
\item{lis}{a list with at least one object to be merge in a single consensus tree. Can be a list of data.frame, distances matrices, matrices or hclust objects.
If the matrices are symmetrical, they will be considered as distance matrices unless the method is "directClustering". The function is expecting that the dataframe have the same number of rows,
that the distance matrices have the same dimensions and that the trees have the same number of leaves. It is expected that the}

\item{method}{the method to use for building the consensus tree. Can be "directClustering", "distanceAverage" or "mergeTrees". If NULL, then the method will be chosen according to the type of objects in lis.}

\item{data_st}{a boolean indicating wether the data need to be standardized (column wise). Used if lis is a list of aSymmetrical dataSets or if "directClustering" is the method.}

\item{dist_st}{a boolean indicating wether the distance matrices need to be standardized before averaging them or performing clustering on them. Apply to methods mergeTrees and "distanceAverage."}

\item{tree_st}{a boolean indicating wether the trees need to be standardized before merging (only apply to mergeTrees method)}

\item{data_center}{a boolean. Used if lis is a list of aSymmetrical dataSets or if "directClustering" is the method.}

\item{distance_method}{distance used to create distance matrices. Default is "euclidean". Not used if objects are trees or distance matrices.}

\item{linkage_method}{linkage criteria used to create the trees. Default is "ward.D2". Not used if objects are trees.}

\item{verbose}{boolean. Display (or not) what the method is doing. Default is FALSE.}
}
\value{
A list of class hclust, being the consensus tree, with the following components: height, merge, method, order, and, if any, labels. For more information about these components, please see hclust help page.
}
\description{
The function assume all of the data.frame or matrices have similar dimensions: for directClustering methods, the datasets are merged using "cbind", it is expected that the matrices have the same number of rows.
The class of the objects in the list and the symmetry of the matrices, are determined by the first object in the list.
}
\section{Warning}{
 There is no check step for the rownames/labels for the data.frame and trees. Please make sure the distance matrices or data.frame in the list are so the rows are ordered in the same ways.
}

\seealso{
\code{\link[stats]{hclust}}, \code{\link[stats]{dist}}, \code{\link{mergeTrees}}, \code{\link[ape]{consensus}}
}
\author{
Audrey Hulot, \email{audrey.hulot@inra.fr}, Julien Chiquet, Guillem Rigaill
}
