% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_trees.R
\name{mergeTrees}
\alias{mergeTrees}
\title{Merge a set of hclust objet into a single tree}
\usage{
mergeTrees(hc.list, standardize = FALSE)
}
\arguments{
\item{hc.list}{a list with a least one hclust object to be merge in a single consensus tree}

\item{standardize}{a boolean indicating wether the heights of the different trees should be normalized before merged}
}
\description{
Merge a set of hclust objet into a single tree
}
